+++ src/EventHandler/CommandType.php
@@ -1,0 +1,16 @@
+<?php declare(strict_types=1);
+
+namespace danog\MadelineProto\EventHandler;
+
+use JsonSerializable;
+
+enum CommandType: string implements JsonSerializable
+{
+    case SLASH = '/';
+    case DOT = '.';
+    case BANG = '!';
+    public function jsonSerialize(): mixed
+    {
+        return $this->value;
+    }
+}