+++ src/EventHandler/SimpleFilter/IsForwarded.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace danog\MadelineProto\EventHandler\SimpleFilter;
+
+/** Allows only forwarded messages */
+interface IsForwarded
+{
+}