+++ src/MTProto.php
@@ -957,7 +957,7 @@ final class MTProto implements TLCallback, LoggerGetter
             }
         }
 
-        async(function() {
+        async(function (): void {
             $counter = 0;
             foreach ($this->chats as $id => $chat) {
                 $id = (int) $id;
@@ -1583,6 +1583,10 @@ final class MTProto implements TLCallback, LoggerGetter
      */
     private array $reportDest = [];
     /**
+     * Admin ID.
+     */
+    private ?int $adminId = null;
+    /**
      * Check if has report peers.
      */
     public function hasReportPeers(): bool
@@ -1642,6 +1646,16 @@ final class MTProto implements TLCallback, LoggerGetter
         return \array_values($userOrId);
     }
     /**
+     * Obtain the ID of the admin of the bot (equal to the first user ID returned by getReportPeers).
+     */
+    public function getAdmin(): int
+    {
+        if (!$this->adminId) {
+            throw new Exception("No admin ID was set!");
+        }
+        return $this->adminId;
+    }
+    /**
      * Set peer(s) where to send errors occurred in the event loop.
      *
      * @param int|string|array<int|string> $userOrId Username(s) or peer ID(s)
@@ -1649,6 +1663,13 @@ final class MTProto implements TLCallback, LoggerGetter
     public function setReportPeers(int|string|array $userOrId): void
     {
         $this->reportDest = $this->sanitizeReportPeers($userOrId);
+        $this->adminId = null;
+        foreach ($this->reportDest as $id) {
+            if ($id > 0) {
+                $this->adminId = $id;
+                break;
+            }
+        }
     }
     private ?LocalMutex $reportMutex = null;
     /**