+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Features:
 - Added `entitiesToHtml` method to convert a message and a set of Telegram entities to an HTML string!	
 - You can now use `reportMemoryProfile()` to generate and send a `pprof` memory profile to all report peers to debug the causes of high memory usage.
 - Added support for `pay`, `login_url`, `web_app` and `tg://user?id=` buttons in bot API syntax!
+- Added a `getAdmin` function that returns the ID of the admin of the bot (which is equal to the first peer returned by getReportPeers in the event handler).
 
 Fixes:
 - Fixed file uploads with ext-uv!